+++ stack/__main__.py
@@ -68,6 +68,7 @@ if __name__ == "__main__":
             "redis-stack-server",
             "redisinsight",
             "redisinsight-web",
+            "redis-tools",
         ],
     )
 
@@ -114,11 +115,13 @@ if __name__ == "__main__":
         from stack.recipes.redis_stack_server import \
             RedisStackServer as pkgklass
     elif opts.PACKAGE == "redis-stack":
-        from stack.recipes.redis_stack import RedisStack as pkgklass
+        from stack.recipes.redis_stack_server import RedisStack as pkgklass
     elif opts.PACKAGE == "redisinsight":
         from stack.recipes.redisinsight import RedisInsight as pkgklass
     elif opts.PACKAGE == "redisinsight-web":
         from stack.recipes.redisinsight import RedisInsightWeb as pkgklass
+    elif opts.PACKAGE == "redis-tools":
+        from stack.recipes.redis_tools import RedisTools as pkgklass
     else:
         sys.stderr.write(f"{opts.PACKAGE} is an unsupported package recipe.")
         sys.exit(3)
